services/ml-models/mediapipe/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
mediapipe==0.10.9
opencv-python-headless==4.8.1.78
numpy==1.26.2
pillow==10.1.0
python-multipart==0.0.6
